ncvisual_from_sixel() atop ncsixel_as_rgba()
